+++ arez/src/main/java/react4j/arez/ReactArezComponent.java
@@ -16,6 +16,7 @@ import arez.annotations.Priority;
 import arez.spy.ObservableValueInfo;
 import arez.spy.ObserverInfo;
 import elemental2.core.JsObject;
+import elemental2.promise.Promise;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Stream;
@@ -43,6 +44,11 @@ import static org.realityforge.braincheck.Guards.*;
 public abstract class ReactArezComponent
   extends Component
 {
+  /**
+   * A non-null lock that will be released in the next micro-task which will schedule any renders required.
+   */
+  @Nullable
+  private static Disposable c_schedulerLock;
   private static int c_nextComponentId = 1;
   private final int _arezComponentId;
   private boolean _renderDepsChanged;
@@ -161,7 +167,36 @@ public abstract class ReactArezComponent
   @Override
   protected final ReactNode performRender()
   {
+    pauseArezSchedulerUntilRenderLoopComplete();
     return Disposable.isDisposed( this ) ? null : trackRender();
+  }
+
+  /**
+   * The first time an arez component is rendered it will lock the Arez scheduler and release
+   * the lock in the micro-task immediately following the task that prompted the render. If this
+   * is not done it is possible that Arez can re-trigger a component render when the scheduler is
+   * triggered after the tracked render completes but before the render method has returned to the
+   * react runtime. This results in error message from react as a setState()/forceRender() was invoked
+   * while still within a render() method.
+   *
+   * <p>NOTE: While render methods are read-only transactions, they can un-observe components with
+   * <code>disposeOnDeactivate=true</code> that would result in the component being disposed and
+   * triggering an update that would mark particular React Components/Observers as STALE and trigger
+   * a re-render of that component.</p>
+   */
+  private void pauseArezSchedulerUntilRenderLoopComplete()
+  {
+    if ( null == c_schedulerLock )
+    {
+      c_schedulerLock = getContext().pauseScheduler();
+      // Use a hack of an empty promise that immediately resolves to
+      // schedule the block immediately after this call stack pops.
+      Promise.resolve( (Object) null ).then( ignored -> {
+        c_schedulerLock.dispose();
+        c_schedulerLock = null;
+        return null;
+      } );
+    }
   }
 
   /**